refactor(protocol): complete device and compatibility boundaries
