+++ packages/devices/src/descriptors/inductor.ts
@@ -1,0 +1,17 @@
+import type { DeviceDescriptor } from "../contract.js";
+
+export const inductorDevice = {
+  id: "inductor",
+  symbolId: "inductor",
+  deviceClass: "inductor",
+  referencePrefix: "L",
+  pinOrder: ["1", "2"],
+  targetPolicy: "builtin",
+  requiredParameters: ["value"],
+  dialects: ["spice", "spectre"],
+  capabilities: {
+    supportsModel: false,
+    supportsBulkBinding: false,
+    supportsValueAnnotation: true,
+  },
+} satisfies DeviceDescriptor;
