+++ packages/devices/src/registry.test.ts
@@ -3,12 +3,14 @@ import { describe, expect, it } from "vitest";
 import {
   builtInDeviceDescriptors,
   deviceDescriptor,
+  deviceDescriptorById,
   validateDeviceDescriptors,
 } from "./index.js";
 
 describe("built-in device registry", () => {
   it("contains internally valid, uniquely identified descriptors", () => {
     expect(validateDeviceDescriptors(builtInDeviceDescriptors)).toEqual([]);
+    expect(deviceDescriptorById("nmos")).toBe(deviceDescriptor("nmos"));
   });
 
   it("preserves MOS electrical and netlist behavior", () => {
@@ -35,5 +37,27 @@ describe("built-in device registry", () => {
       pinOrder: ["P"],
       targetPolicy: "none",
     });
+  });
+
+  it("rejects descriptor capability claims that would change device meaning", () => {
+    const nmos = deviceDescriptor("nmos");
+    expect(nmos).toBeDefined();
+    if (!nmos) return;
+    expect(
+      validateDeviceDescriptors([
+        {
+          ...nmos,
+          id: "invalid-bulk-device",
+          symbolId: "invalid-bulk-device",
+          deviceClass: "resistor",
+        },
+      ]),
+    ).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          message: "Only MOS devices may support bulk binding",
+        }),
+      ]),
+    );
   });
 });
