+++ packages/model/src/schema/index.ts
@@ -2,8 +2,12 @@ export * from "./common.js";
 export * from "./source.js";
 export * from "./instance.js";
 export * from "./connectivity.js";
-export * from "./annotation.js";
+export * from "./routing.js";
+export * from "./rich-text.js";
+export * from "./annotations.js";
 export * from "./drafting.js";
+export * from "./presentation.js";
+export * from "./validation.js";
 export * from "./document.js";
 export * from "./project.js";
 export * from "./types.js";
