+++ packages/model/src/schema/presentation.ts
@@ -1,0 +1,44 @@
+import { z } from "zod";
+
+import { StableIdSchema } from "./common.js";
+
+export const PresentationIntentSchema = z.strictObject({
+  styleProfileId: StableIdSchema,
+  grid: z.number().int().positive(),
+  compactness: z.enum(["loose", "normal", "compact"]),
+  flow: z
+    .strictObject({
+      power: z.literal("top").optional(),
+      ground: z.literal("bottom").optional(),
+      input: z.literal("left").optional(),
+      output: z.literal("right").optional(),
+    })
+    .optional(),
+});
+export const MosBulkDefaultsSchema = z.strictObject({
+  nmosNetId: StableIdSchema.optional(),
+  pmosNetId: StableIdSchema.optional(),
+});
+export const LayoutGroupSchema = z.strictObject({
+  id: StableIdSchema,
+  kind: z.enum([
+    "differential-pair",
+    "current-mirror",
+    "matched-pair",
+    "custom",
+  ]),
+  objectIds: z.array(StableIdSchema).min(1),
+  locked: z.boolean(),
+});
+export const LayoutConstraintSchema = z.strictObject({
+  id: StableIdSchema,
+  kind: z.enum([
+    "align-x",
+    "align-y",
+    "symmetric",
+    "equal-spacing",
+    "keep-clear",
+  ]),
+  objectIds: z.array(StableIdSchema).min(2),
+  locked: z.boolean(),
+});
