+++ packages/model/src/schema/source.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 
 import { StableIdSchema } from "./common.js";
+import { reportDuplicateIds } from "./validation.js";
 export const SourcePositionSchema = z.strictObject({
   offset: z.number().int().nonnegative(),
   line: z.number().int().positive(),
@@ -35,17 +36,7 @@ export const SourceManifestSchema = z
     files: z.array(SourceFileRecordSchema),
   })
   .superRefine((manifest, context) => {
-    const seen = new Set<string>();
-    for (const [index, file] of manifest.files.entries()) {
-      if (seen.has(file.id)) {
-        context.addIssue({
-          code: "custom",
-          message: `Duplicate ID: ${file.id}`,
-          path: ["files", index, "id"],
-        });
-      }
-      seen.add(file.id);
-    }
+    reportDuplicateIds(manifest.files, "files", context);
   });
 export const SymbolLibraryLockSchema = z.strictObject({
   id: StableIdSchema,
