+++ packages/project-protocol/src/index.ts
@@ -1,1 +1,6 @@
-export * from "./protocol.js";
+export * from "./diagnostics.js";
+export * from "./load.js";
+export * from "./previous-to-current.js";
+export * from "./save.js";
+export * from "./storage.js";
+export * from "./version.js";
