+++ packages/project-protocol/src/previous-to-current.ts
@@ -1,0 +1,1 @@
+export { upgradePreviousProject } from "./transforms/project.js";
