+++ packages/project-protocol/src/version.ts
@@ -1,0 +1,5 @@
+/**
+ * The rolling compatibility window is deliberately explicit. Advancing the
+ * current schema replaces this adapter instead of extending a migration chain.
+ */
+export const PREVIOUS_PROJECT_SCHEMA_VERSION = 10;
