+++ packages/symbols/package.json
@@ -14,8 +14,10 @@
     "build": "tsc -p tsconfig.json"
   },
   "dependencies": {
-    "@icm/devices": "workspace:*",
     "@icm/model": "workspace:*",
     "zod": "4.4.3"
+  },
+  "devDependencies": {
+    "@icm/devices": "workspace:*"
   }
 }
